refactor: use context instead of internal header
